requirements.txt

fastapi
uvicorn
pymupdf
ezdxf
reportlab
python-multipart
